Remove config from upstream remote for checking out github PRs
It turns out that this just polluted the output of `git adog`
to the point where it became useless. Being able to view the
commit history is more important than being able to check out
PRs easily. Besides, `gh pr checkout` can be used for this.
And PRs are a github-specific thing, so it makes sense that
one would use the github CLI for checking them out.